Cleans client (#8)
* borra archivos innecesarios de react. agrega readme al proyecto.

* tambien normaliza el uso de .jsx

* normaliza nombres de componentes y carpetas
