+++ README.md
@@ -1,0 +1,19 @@
+<h1>Readme del proyecto:</h1>
+
+<h2>Integrantes:</h2>
+
+- <a href='https://github.com/EzeTP'>Ezequiel Tejada</a>
+
+- <a href='https://github.com/JavierMusso'>Javier Musso</a>
+
+- <a href='https://github.com/jonrovner'>Jonathan Rovner</a>
+
+- <a href='https://github.com/'>Jose Puello</a>
+
+- <a href='https://github.com/leo10-kz'>Leonel Quiroga</a>
+
+- <a href='https://github.com/LucasZupan'>Lucas Zupan</a>
+
+- <a href='https://github.com/martinjuncos'>Martin Juncos</a>
+
+- <a href='https://github.com/NicoYanez9621'>Nicolas Yañez</a>
