--- client/src/App.js
@@ -1,12 +1,0 @@
-import "./App.css";
-import Ownersignup from "./components/ownerSignUp/OwnerSignUp";
-
-function App() {
-  return (
-    <div className="App">
-      <Ownersignup />
-    </div>
-  );
-}
-
-export default App;
