+++ client/src/App.jsx
@@ -1,0 +1,11 @@
+import Ownersignup from "./components/OwnerSignUp/OwnerSignUp";
+
+function App() {
+  return (
+    <div>
+      <Ownersignup />
+    </div>
+  );
+}
+
+export default App;
