+++ client/src/components/ownerSignUp/OwnerSignUp.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import style from "./OwnerSignUp.module.scss";
+
 const Ownersignup = () => {
-  return <div className={style.contenedor}>Owner Sign up</div>;
+  return <div>Owner Sign up</div>;
 };
 
 export default Ownersignup;
