--- client/src/components/ownerSignUp/OwnerSignUp.module.scss
@@ -1,2 +1,0 @@
-.contenedor {
-}
